+++ views/default/index.php
@@ -65,7 +65,7 @@ if (isset($this->context->module->panels['db']) && isset($this->context->module-
             [
                 'attribute' => 'time',
                 'value' => function ($data) {
-                    return '<span class="nowrap">' . Yii::$app->formatter->asDatetime($data['time'], 'short') . '</span>';
+                    return '<span class="nowrap">' . Yii::$app->formatter->asDatetime($data['time'], 'yyyy-MM-dd HH:mm:ss') . '</span>';
                 },
                 'format' => 'html',
             ],
@@ -111,6 +111,21 @@ if (isset($this->context->module->panels['db']) && isset($this->context->module-
             ],
             [
                 'attribute' => 'statusCode',
+                'value' => function ($data) {
+                    $statusCode = $data['statusCode'];
+                    if ($statusCode === null) {
+                        $statusCode = 200;
+                    }
+                    if ($statusCode >= 200 && $statusCode < 300) {
+                        $class = 'label-success';
+                    } elseif ($statusCode >= 300 && $statusCode < 400) {
+                        $class = 'label-info';
+                    } else {
+                        $class = 'label-important';
+                    }
+                    return "<span class=\"label {$class}\">$statusCode</span>";
+                },
+                'format' => 'raw',
                 'filter' => $statusCodes,
                 'label' => 'Status code'
             ],