+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Yii Framework 2 debug extension Change Log
 - Enh #16: Added ability to EXPLAIN queries in Database panel for MySQL, SQLite, PostgreSQL and Cubrid (laszlovl, samdark)
 - Enh #19: Mark selected log item in dropdown list with bold font and an arrow (idMolotov)
 - Enh: More compact toolbar (samdark)
+- Enh: Display colorful status at index page (samdark)
+- Enh: More readable format for date and time at index page (samdark)
 
 
 2.0.4 May 10, 2015